requirements.txt
google-genai>=0.5.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
openpyxl>=3.1.0
rich>=13.0.0
python-dotenv>=1.0.0
